routes protected...authMiddleware, CRUD complete

+++ backend/controllers/goalController.js
@@ -2,12 +2,13 @@ const asyncHandler = require('express-async-handler')
 
 // Import required model
 const Goal = require('../models/goalModel')
+const User = require('../models/userModels')
 
 // @desc    Get goals
 // @route   GET  /api/goals
 // @access  private
 const getGoals = asyncHandler( async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -22,7 +23,8 @@ const setGoal = asyncHandler( async (req, res) => {
     }
     // create model
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     // send model
@@ -38,6 +40,20 @@ const updateGoal = asyncHandler( async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -58,6 +74,20 @@ const deleteGoal = asyncHandler( async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // Make sure logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     // delete goal
     await Goal.deleteMany(goal)
 
@@ -68,5 +98,5 @@ module.exports = {
     getGoals,
     setGoal,
     updateGoal,
-    deleteGoal
+    deleteGoal,
 }
